Fixed Power function and epsilon
I was doing powers incorrectly. I had assumed it was a^x form when in
reality it is a ** x. I learned this while stepping through debug.

I also changed epsilon because when running the tests it was giving me
results that were very close but not quite close enough for the number
of significant figures assigned in the unit tests. (found in p tests)