Fixed setsightingfile green light test
So I decided to arrange my project so that it used a resource file for
the sightings files and stuff, so I have modified the test code to
match. This will make so it doesn't work on your side even though it
does on mine. I might change this later.